CU-8u0gnd - Refine the subscription module by the test cases.

+++ src/main/java/org/elastos/hive/subscription/VaultSubscribeResponse.java
@@ -10,7 +10,8 @@ public class VaultSubscribeResponse {
     @SerializedName("quota")
     private int quota;
     @SerializedName("created")
-    private long created;
+    //TODO:
+    private double created;
     @SerializedName("updated")
     private long updated;
 
@@ -26,7 +27,7 @@ public class VaultSubscribeResponse {
         return quota;
     }
 
-    public long getCreated() {
+    public double getCreated() {
         return created;
     }
